requirements.txt
aenum==3.1.11
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
awswrangler==2.15.1
backoff==1.11.1
beautifulsoup4==4.11.1
boto3==1.23.5
botocore==1.26.5
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
dacite==1.6.0
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
fastapi==0.78.0
Flask==2.1.1
frozenlist==1.3.0
gremlinpython==3.6.0
h11==0.13.0
idna==3.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
jsonpath-ng==1.5.3
lxml==4.8.0
MarkupSafe==2.1.1
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.22.3
openpyxl==3.0.10
opensearch-py==1.1.0
packaging==21.3
pandas==1.4.2
pandas-redshift==2.0.5
pg8000==1.29.1
pipreqs==0.4.11
ply==3.11
progressbar2==4.0.0
psycopg2-binary==2.9.3
pyarrow==7.0.0
pycodestyle==2.8.0
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-utils==3.2.3
pytz==2022.1
PyYAML==6.0
redshift-connector==2.0.907
requests==2.27.1
requests-aws4auth==1.1.2
s3transfer==0.5.2
scramp==1.4.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.1
yarg==0.1.9
yarl==1.7.2
